Create const keyword 2 (constant object) in internal js.html
This HTML document showcases how constant objects work in JavaScript. The key point is that while you cannot reassign a const object (e.g., obj = { name: "cow" } would throw an error), you can modify its properties.

In the example:

The constant object obj is initially created with properties animal set to "Lion" and color set to "Yellow".

The animal property is then updated to "Tiger".

A new property legs with the value 4 is added to the object.

The updated object is printed using JSON.stringify().

The const keyword prevents reassigning the object itself but allows modification of its internal properties.