Create converting date to string or number in internal js.html
This HTML document demonstrates converting a Date object into both a numeric and string representation in JavaScript. The Date() object is created to represent the current date and time. The getTime() method is used to convert the date into a numeric value, which represents the number of milliseconds since January 1, 1970, UTC. This numeric value is displayed as "The Numeric date is: [timestamp]". The toString() method is then called to convert the Date object into a readable string format, and it is displayed as "The string date is: [formatted date]". Both values are printed on the webpage.