Create implicit type conversion (String to Number) in internal js.html
This HTML code demonstrates implicit conversion of values to numbers in JavaScript. The first operation ('100' / 50) divides a string by a number, resulting in 2 due to implicit conversion. The second operation ('100' - '50') subtracts two strings, resulting in 50 because strings are converted to numbers for subtraction. The third operation ('100' * true) multiplies a string by a boolean, resulting in 100 because true is implicitly converted to 1. The fourth operation ('100' - false) subtracts a boolean from a string, resulting in 100 because false is implicitly converted to 0. The last operation ('tp' / 50) involves dividing a non-numeric string by a number, resulting in NaN because the string cannot be converted to a valid number.