Update rest parameters in internal js.html
This HTML code demonstrates the use of rest parameters in JavaScript. The sum function uses the ...nums syntax, which collects all the passed arguments into an array called nums. The function then calculates the sum of all numbers by iterating over the nums array.

The sum function is called three times:

The first call sums 1 + 5 + 8 + 20 + 23 + 45, which equals 102.

The second call sums 10 + 20 + 30, which equals 60.

The third call with no arguments results in a sum of 0.