Create extending the abstract class in internal js.html
This code demonstrates the concept of abstract classes and inheritance in JavaScript.

Abstract Class (fruit): The fruit function acts as an abstract class. It has a property name and a method getName. If someone tries to instantiate fruit directly, it throws an error using throw new Error(...), ensuring that fruit cannot be instantiated.

Child Class (Apple): The Apple function is a child class that inherits from fruit. It overrides the name property. The Apple.prototype = Object.create(fruit.prototype) line links Apple to the fruit prototype, allowing Apple to access the getName method.

Creating and Using an Instance: An instance of Apple is created with new Apple("Apple"). It then calls the getName method to get the name of the fruit and displays it on the web page.