Create abstraction in internal js.html
The provided code demonstrates the concept of preventing the instantiation of an abstract class in JavaScript.

Abstract Class (fruit): The fruit function is designed as an abstract class. The constructor checks if it's being instantiated directly (this.constructor === fruit). If so, it throws an error to prevent creating an instance of the fruit class itself.

Error Handling: The try...catch block is used to catch the error thrown when trying to instantiate fruit. The error message is displayed on the web page in the div element with the ID output.

When the code runs, the error message "You can't create the instance of the fruit." will be displayed on the page, because the fruit class cannot be instantiated directly.