Create super keyword 1 in internal js.html
In this code, the Bike class is the parent class, which takes the number of gears as a parameter during initialization. The suzuki class is a child class that extends Bike and also takes a brand and gears as parameters, passing the gears to the parent class using super(gears).

The suzukiBike object is created with the brand "Suzuki" and 4 gears. The brand and gears values are then displayed in the output1 and output2 elements.