Create hierarchical inheritance in internal js.html
This code defines two child classes Honda and Suzuki that extend the parent Bike class. The Bike class is initialized with the gears parameter, while Honda and Suzuki classes extend it with additional properties like model and color.

The objects h_Bike and s_Bike are created for Honda and Suzuki bikes, respectively. The JSON.stringify() method is used to display the properties of these objects in the HTML elements output1 and output2.