Create super keyword 2 in internal js.html
The HTML code defines a simple webpage displaying information about a fashion designer. It uses JavaScript classes to represent a Person and a FashionDesigner class (which extends Person). The FashionDesigner class includes additional methods related to the profession of being a fashion designer.

When the page loads, it creates an instance of FashionDesigner with the name "Sayan" and age 30, then displays:

The profession of the character.

The status of being at home.

Tasks related to being at work and being a fashion designer.