Update inheritance in internal js.html
The HTML code demonstrates the concept of code reusability through inheritance in JavaScript. In this example:

The Bike class is the parent class with properties like gear and a method getGears() that returns the value of gear.

The suzuki class is the child class that inherits from Bike using extends. It has a new property brand and a method getBrand() that returns the brand name.

The super() keyword is used in the child class to call the parent class constructor and inherit the properties and methods.