Create define property method in internal js.html
The given HTML and JavaScript code demonstrates the use of Object.defineProperty() in JavaScript, which adds a new property named gears with a value of 6 to the car object. This property is configured to be writable, enumerable, and configurable. The complete car object, including the new gears property, is then converted to a JSON string and displayed on the web page.