Create bitwise (and) operator in internal js.html
This code demonstrates the bitwise AND (&) operation in JavaScript. The bitwise AND operator compares each bit of the two numbers (a = 5 and b = 7) and returns a new number whose bits are set to 1 only if the corresponding bits of both operands are also 1. In binary, 5 is 0101 and 7 is 0111. Performing 0101 & 0111 results in 0101, which is 5 in decimal. So, the output displayed in the HTML element with ID output will be: a & b = 5.