Create not equal to operator in internal js.html
This HTML page demonstrates the JavaScript !== (strict not equal) comparison operator. It assigns the value 5 to the variable x and compares it with 8 using (x !== 8). Since 5 is not equal to 8, and both are of the same type (number), the comparison evaluates to true, which is then displayed in the paragraph with the ID demo.