Create triple equals to operator in internal js.html
This HTML page demonstrates the JavaScript === (strict equality) operator. It assigns the value 5 to the variable x and compares it with 5 using (x === 5). Since both the value and type are the same (number), the result of the comparison is true, which is then displayed in the paragraph with the ID demo.