Create assignment operator - binary operator in internal js.html
This HTML code uses JavaScript to declare two variables, x and y, where x is assigned the value 5 and y is computed as x + 10, resulting in 15. The values of x and y are then displayed within the div element with the ID output as:
Value of x : 5
Value of y : 15.