Create increment operator (++) in internal js.html
This HTML code demonstrates the use of the ++ operator in JavaScript, which increments the value of x by 1. Initially, x is set to 5, and then x++ increments it by 1, changing x to 6. The value of x after the increment is assigned to the variable z, and z is then displayed in the HTML.

The output will be:

6