Create nested if-else statement in internal js.html
The provided HTML and JavaScript code uses conditional statements (if, else if, else) to check the value of the book variable and display a corresponding message. With the current value of book set to "maths", the output displayed will be "Maths Book". The code checks if book equals "history", "maths", or "economics", and displays respective messages for each. If none of these conditions are met, it displays "Unknown Book". To test different outcomes, you can change the value of book and observe the updated result in the output.