Create if-else statement in internal js.html
Your code correctly checks if the variable age is greater than 18 and displays a message based on the result. When age is set to 15, the message "Does not qualify for driving" will be displayed in the output div, as the condition age > 18 is false. If the value of age is changed to a number greater than 18, like 20, the message will update to "Qualifies for driving". The logic and output are correct, and the condition works as expected.