Create object constructor in internal js.html
This HTML page uses JavaScript to demonstrate an object constructor function named Tree, which initializes objects with name and age properties. When the Tree function is called with the new keyword, it creates an object (tree1) with name set to "palm" and age set to 5. The script then accesses these properties and displays them inside the paragraph element with the ID output, resulting in the text: “name = palm, age = 5” being shown on the web page.