Update getter and setter methods for improving data quality and providing security in internal js.html
This HTML page demonstrates the use of Object.defineProperty() to define getter and setter methods for an object (door). Initially, the door object has a property size set to 20. A getter method getSize is defined to return the value of size, and a setter method setSize is defined to update the value of size. The script first displays the door's size as 20, then updates the size to 30 using the setter and displays the updated size.