Create boolean function in internal js.html
This HTML page demonstrates the use of JavaScript’s Boolean function and comparison operators. It evaluates three expressions: whether 100 is greater than 90 (which is true), whether 100 is less than 90 (which is false), and whether 100 is equal to 90 using == (also false). However, each result is written to the same paragraph element (output) using innerHTML, which overwrites the previous content each time. As a result, only the last line "100 == 90 : false" will appear on the page. To display all results, += should be used instead of = for innerHTML.