Create static method with normal method in internal js.html
This HTML page defines a JavaScript class Num with a static method getSqr(a) that returns the square of a number a, and a regular method printSqr(a) that prints the result using the static method. An instance num1 of Num is created, and printSqr(6) is called, which calculates the square of 6 using the static method and displays the result in the paragraph with ID demo.