Create setter method in internal js.html
This HTML page defines a JavaScript object wall with a color property and a setter method setColor to update its color. Initially, the color is "brown", which is displayed in the paragraph with ID output. Then, the setter is used to update the color to "Red" by assigning wall.setColor = "Red". The updated color is then displayed.