Create static method with constructor in internal js.html
This HTML page defines a JavaScript class Num with a constructor that assigns the result of a static method getSqrt(a) to the instance property sqrt. Despite the method name getSqrt, it actually returns the square of a (i.e., a * a), not the square root. An instance num1 is created with the value 10, so 10 * 10 = 100 is assigned to num1.sqrt and displayed in the paragraph with ID output.