Create single static method in internal js.html
This HTML page defines a JavaScript class Table with a static method printSize() that returns the size of the table, and a regular method getColor() that returns its color. The static method is called directly on the class using Table.printSize() and displayed first. Then, an instance tableObj of the Table class is created, and its getColor() method is called to get the table's color.