Create function call 2 in internal js.html
This HTML page demonstrates how to use the JavaScript call() method to invoke a function with a specified this context. The function printMessage returns a string containing the name and age of the context it's called with. Two objects, person1 and person2, are defined with different age values. Using printMessage.call(person1) and printMessage.call(person2), the function is executed with this bound to each object. The results are displayed in the paragraphs with IDs output1 and output2.