Create function call 1 in internal js.html
This HTML page demonstrates the use of the JavaScript call() method, which allows you to invoke a function with a specific this context and pass additional arguments. The function printSum(p1, p2) calculates the sum of the properties num1 and num2 from the nums object, along with the two arguments p1 and p2.

In this case, the object nums has properties num1 = 5 and num2 = 10. The call() method is used to call printSum with nums as the this context, and it passes 40 and 32 as arguments. The result of the sum is displayed in the paragraph with ID output.