Create this keyword in internal js.html
This HTML page demonstrates the use of the this keyword within a JavaScript constructor function. The function Car(name, model, year) is a constructor that assigns name, model, and year properties to an object. When a new instance of Car is created with new Car("OD", "Q6", 2020), the this keyword is used to assign these values to the new object ODCar. The JSON.stringify() method is then used to convert the ODCar object into a string representation and display it in the paragraph with ID output.