Create arrow function with ternary operator in internal js.html
This HTML page demonstrates the use of an arrow function combined with a ternary operator in JavaScript. The variable isMul is set to true, which makes the first arrow function execute, performing a multiplication operation (5 * 5). If isMul were false, the second arrow function would execute, performing an addition operation (5 + 5). The function is immediately invoked using func().

Since isMul is true, the multiplication function is called, and the result (25) is displayed in the paragraph with ID output.