Create arrow or lambda function 1 in internal js.html
This HTML page demonstrates a simplified arrow function (also known as a lambda function) in JavaScript. The divide function is defined using an arrow function that takes two parameters x and y and returns the result of x / y.

The function is called with arguments 10 and 5, and the result (2) is displayed in the div with the ID output.