Create parameters binding of same function 2 in internal js.html
This HTML page demonstrates the bind() method in JavaScript. The bind() method is used to create a new function with a specified this value and preset arguments.

In this example, the multiply function, which takes two parameters (x and y), is bound to the value 2 as the first argument using bind(). This creates a new function multiplyByTwo, where the first parameter is fixed to 2, and you can pass the second parameter when calling the function.

When multiplyByTwo(5) is invoked, the result is 2 * 5, which is 10. This result is then displayed in the div with ID output.