Create functions binding 1 in internal js.html
In this HTML and JavaScript code, the greet function is defined to take a message parameter and display it along with the name property of the object this. The greet.call(person, 'Hello') method is used to invoke the greet function, explicitly setting this to the person object, which has a name property of 'John Doe'. The result is that the message "Hello, John Doe" is displayed inside the <div id="demo"> element. The call() method allows invoking a function with a specified this context, in this case, the person object.