Create fucntion with no arguments but a return value in internal js.html
In this HTML and JavaScript code, a new function func is created using the Function constructor. This function is defined to return the result of the expression 10 + 20. The func() is called, and its result (which is 30) is displayed in the <p id="output"> element. The result is shown in the format: "The value returned from the function is: 30". The Function constructor allows for dynamic function creation, where the body of the function can be passed as a string.