Create function with arguments and a return value in internal js.html
In this HTML and JavaScript code, a function func is dynamically created using the Function constructor. This function takes two parameters, p and q, and returns their product (p * q). The function is then invoked with the arguments 5 and 7, which results in the value 35. This value is displayed in the <p id="output"> element in the following format: "The value returned from the function is: 35".

The Function constructor allows you to define functions as strings, which can be useful in cases where the function body is generated dynamically or needs to be specified at runtime.