Create call back function with events in internal js.html
In this HTML and JavaScript code, a button with the text "Click Me" is provided, and when clicked, it triggers a function that displays a message in the <p id="output"> element.

The button.addEventListener('click', function () {...}) line adds an event listener to the button. When the button is clicked, the anonymous function is executed.

Inside this function, the output.innerHTML is updated to display the message 'You have clicked the button.', followed by a line break <br>.

So, when the user clicks the button, the message "You have clicked the button." will appear in the <p id="output"> element below the button.