Create weak map object 2 in internal js.html
This HTML and JavaScript code creates a Map with string keys and number-word values, then iterates over it using a for...of loop. Each map entry is a key-value pair represented as an array (e.g., ["1", "one"]), which, when appended to innerHTML, gets automatically converted to a comma-separated string like "1,one". As a result, the output displays each pair on a new line, showing the map’s contents in the order they were inserted.