Create iterate with map method in internal js.html
This HTML and JavaScript code uses the map() method to iterate over the array [true, false, 50, 40, "Hi"]. For each element in the array, the code appends the element followed by a comma to the innerHTML of the <p id="demo"> element, resulting in the display of the array elements separated by commas. While the map() method returns a new array, it is not used here; instead, it serves to iterate over the array and perform the action of displaying the elements in the HTML. The final output displayed is: "Using the map() method: true, false, 50, 40, Hi, ".