Create promise in internal js.html
This HTML page defines a promise1 object using JavaScript's Promise constructor, which checks if the variable num equals 10. Since num is indeed 10, the promise resolves with the message "The value of the number is 10 <br>". However, instead of handling the resolved value with .then(), the code directly appends the promise1 object to the <div> with ID output, which results in displaying [object Promise] rather than the actual message.