Create object entries method in internal js.html
This HTML page defines a numbers object with numeric key-value pairs and uses Object.entries() to convert the object into an array of [key, value] pairs. It then uses a for...of loop to iterate over these pairs, appending each key and its corresponding value to the paragraph with ID output. The output is displayed in the format num1: 10, num2: 20, etc., each on a new line using <br>.