Create display objects in internal js.html
This HTML page defines a fruit object with properties name and price, and accesses these properties using two different methods: dot notation (fruit.name) and bracket notation (fruit["price"]). It then combines these values into a message and sets it as the content of the paragraph with ID output, displaying: "The price of the Banana is: 100".