Create with keyword in internal js.html
This HTML page defines a Book constructor function that creates book objects with title, author, price, and a method addPrice. The addPrice function uses the with statement to set the price property of the book object. A book instance myBook is created with the title "Perl" and author "Mohtashim", and its price is set to 100 using the addPrice method. The output <div> is then updated to display the book's title, author, and price on separate lines using <br>.