Create object in internal js.html
This HTML page creates a book object using the new Object() constructor and assigns it two properties: subject with the value "Perl" and author with the value "Mohtashim". It then updates the content of the paragraph with ID output to display the book's name and author using string concatenation and line breaks (<br>), resulting in the message: "Book name is : Perl" and "Book author is : Mohtashim".